display visual view chart
