--- frontend/src/features/visualView/mockGraphData.json
@@ -1,82 +1,0 @@
-[
-  [
-    "Dīghanikāya (dn)",
-    "Dīghanikāya_(dn)",
-    2115384
-  ],
-  [
-    "Dīghanikāya (dn)",
-    "Majjhimanikāya_(mn)",
-    3385820
-  ],
-  [
-    "Dīghanikāya (dn)",
-    "Saṁyuttanikāya_(sn)",
-    2773096
-  ],
-  [
-    "Dīghanikāya (dn)",
-    "Aṅguttaranikāya_(an)",
-    3012780
-  ],
-  [
-    "Majjhimanikāya (mn)",
-    "Dīghanikāya_(dn)",
-    2727718
-  ],
-  [
-    "Majjhimanikāya (mn)",
-    "Majjhimanikāya_(mn)",
-    7802731
-  ],
-  [
-    "Majjhimanikāya (mn)",
-    "Saṁyuttanikāya_(sn)",
-    7195361
-  ],
-  [
-    "Majjhimanikāya (mn)",
-    "Aṅguttaranikāya_(an)",
-    6260930
-  ],
-  [
-    "Saṁyuttanikāya (sn)",
-    "Dīghanikāya_(dn)",
-    2027785
-  ],
-  [
-    "Saṁyuttanikāya (sn)",
-    "Majjhimanikāya_(mn)",
-    6147539
-  ],
-  [
-    "Saṁyuttanikāya (sn)",
-    "Saṁyuttanikāya_(sn)",
-    13911018
-  ],
-  [
-    "Saṁyuttanikāya (sn)",
-    "Aṅguttaranikāya_(an)",
-    6816528
-  ],
-  [
-    "Aṅguttaranikāya (an)",
-    "Dīghanikāya_(dn)",
-    2661690
-  ],
-  [
-    "Aṅguttaranikāya (an)",
-    "Majjhimanikāya_(mn)",
-    6873170
-  ],
-  [
-    "Aṅguttaranikāya (an)",
-    "Saṁyuttanikāya_(sn)",
-    8363097
-  ],
-  [
-    "Aṅguttaranikāya (an)",
-    "Aṅguttaranikāya_(an)",
-    11861895
-  ]
-]
