+++ frontend/src/utils/api/dbApi.ts
@@ -15,7 +15,10 @@ import { getAvailableLanguages } from "./endpoints/utils/available-languages";
 import { getCountMatches } from "./endpoints/utils/count-matches";
 import { getTextDisplayName } from "./endpoints/utils/displayname";
 import { getFolios } from "./endpoints/utils/folios";
-import { getVisualViewCollections } from "./endpoints/visual-view";
+import {
+  getVisualGraphData,
+  getVisualViewCollections,
+} from "./endpoints/visual-view";
 
 export const DbApi = {
   //* VIEWS
@@ -101,5 +104,17 @@ export const DbApi = {
   VisualViewCollections: {
     makeQueryKey: (language: DbLanguage) => ["visualViewCollections", language],
     call: (language: DbLanguage) => getVisualViewCollections(language),
+  },
+  VisualViewGraphData: {
+    makeQueryKey: (
+      language: DbLanguage,
+      inquiryCollection: string,
+      hitCollections: string[],
+    ) => ["visualViewGraphData", language, inquiryCollection, hitCollections],
+    call: (
+      language: DbLanguage,
+      inquiryCollection: string,
+      hitCollections: string[],
+    ) => getVisualGraphData(language, inquiryCollection, hitCollections),
   },
 };
